fixed handling of more than ~30 generations depth. added option to control the maximal amount of generations parsed.

+++ KekuleHtml/Models/FamilyTree.cs
@@ -26,7 +26,7 @@ namespace KekuleHtml.Models
             }
         }
 
-        public Person? GetPerson(int kekuleNumber)
+        public Person? GetPerson(ulong kekuleNumber)
         {
             return AllPersons.FirstOrDefault(p => p.KekuleNumber == kekuleNumber);
         }
